+++ src/events/Ready.js
@@ -9,5 +9,6 @@ module.exports = class extends Event {
 
   run() {
     this.client.events.forEach(e => e._init());
+    this.client.commands.forEach(e => e._init());
   }
 };