+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## [0.3.0] - 2019-08-13
+
+### Added
+
+- Core Option: db.
+- Commands.load() method.
+- Events.load() method.
+
+### Changed
+
+### Fixed
+
+- Command.init() method.
+
 ## [0.2.7] - 2019-08-12
 
 ### Added
@@ -8,7 +22,7 @@
 
 ### Fixed
 
-PermissionLevels is not a constructor error.
+- PermissionLevels is not a constructor error.
 
 ## [0.2.6] - 2019-08-12
 
@@ -16,8 +30,8 @@ PermissionLevels is not a constructor error.
 
 - id option to commands.
 - id option to events.
-- CoreSetting: ignoreSelf.
-- CoreSetting: ignoreBots.
+- Core Option: ignoreSelf.
+- Core Option: ignoreBots.
 - Usage for commands.
 - CHANGELOG.md file.
 - Command.reload() method.